# public/openapi.yaml
openapi: 3.1.0
info:
  title: mindlog · id — API
  version: 0.2.2
  description: |
    Carte d'identité en ligne (1 page + API + MCP).

    ## Authentification
    Deux mécanismes, résolus par `currentIdentity()` :
    1. **Cookie de session** `mindlog_session` (HttpOnly, Secure, SameSite=Lax) —
       posé après login passkey ou échange clé→session. Préféré pour le navigateur.
       Pour les méthodes non sûres (POST/PUT/PATCH/DELETE), l'en-tête `Origin` est
       vérifié quand il est présent (protection CSRF).
    2. **Clé d'accès** — secret de 24 octets (base64url). Transmis soit par l'en-tête
       `X-Access-Key`, soit par le paramètre de requête `?key=...`. Utilisé par les
       liens privés `/k/{clé}`, le serveur MCP et le flux SSE.

    Les routes « contacts » exigent une relation **réciproque** entre les deux handles.
    Les routes « publiques » n'exigent rien (un viewer authentifié élargit la réponse).

    ## Sécurité transverse
    - **Rate-limiting** (par IP) : `429 Too Many Requests` + `Retry-After`.
      Backstop global de 300 req/min sur `/api/*` ; limites strictes sur
      `POST /api/identities` (10/h), `POST /api/recover` (5/15 min),
      `POST /api/passkeys/auth/*` (20/15 min), `GET /api/search` (60/min).
    - **Taille du corps** : `413` au-delà de 256 Ko (JSON) ou 55 Mo (multipart).
    - **CSRF** : pour une mutation authentifiée par cookie, l'`Origin` (ou à
      défaut le `Referer`) doit correspondre à l'origine de l'app.
...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mindlog_session
    accessKeyHeader:
      type: apiKey
      in: header
      name: X-Access-Key
    accessKeyQuery:
      type: apiKey
      in: query
      name: key
    oauth2:
      type: oauth2
      description: OAuth 2.1 (Authorization Code + PKCE S256, DCR). mindlog est AS + RS.
      flows:
        authorizationCode:
          authorizationUrl: https://id.mindlog.today/oauth/authorize
          tokenUrl: https://id.mindlog.today/oauth/token
          refreshUrl: https://id.mindlog.today/oauth/token
          scopes:
            "mindlog:identity": Gérer la carte d'identité du compte connecté
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Ok:
      type: object
      properties:
        ok: { type: boolean, example: true }
...
security:
  - sessionCookie: []
  - accessKeyHeader: []
  - accessKeyQuery: []
paths:
  /api/identities:
    post:
      tags: [Identités]
      summary: Créer une identité (handle + clé d'accès)
      security: []
      description: Protégé par Cloudflare Turnstile en production (sauf localhost).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [handle]
              properties:
                handle: { type: string, pattern: '^[a-z0-9][a-z0-9_-]{1,29}$' }
                display_name: { type: string }
                email: { type: string, format: email }
                autoCreated: { type: boolean }
                turnstileToken: { type: string }
      responses:
        '201':
          description: Identité créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  handle: { type: string }
                  accessKey: { type: string }
                  publicUrl: { type: string }
                  privateUrl: { type: string }
        '400': { description: Handle invalide ou Turnstile échoué, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Handle déjà pris ou réservé }
  /api/recover:
    post:
      tags: [Identités]
      summary: Récupérer un compte par email (régénère la clé d'accès)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [handle, email]
              properties:
                handle: { type: string }
                email: { type: string, format: email }
      responses:
        '200':
          description: |
            Si l'email correspond : nouvelle clé envoyée par mail (`sent: true`).
            Si le mail n'est pas configuré ou échoue, la clé est renvoyée en clair.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
                  handle: { type: string }
                  accessKey: { type: string }
                  privateUrl: { type: string }
        '400': { description: Handle/email ne correspondent à aucun compte }
  /api/search:
    get:
      tags: [Identités]
      summary: Rechercher des identités par handle ou nom
      security: []
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Résultats (max 12)
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        handle: { type: string }
                        display_name: { type: string }
                        title: { type: string }
                        has_photo: { type: boolean }
...
  /api/me:
    get:
      tags: [Identités]
      summary: Carte privée complète de l'identité authentifiée
      responses:
        '200': { description: Données complètes (champs, événements, relations, notifications, requests, settings…) }
        '401': { description: Non authentifié }
    delete:
      tags: [Identités]
      summary: Suppression définitive du compte (RGPD)
      responses:
        '200': { description: Supprimé, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        '401': { description: Non authentifié }
...
  /api/auth/redeem-pin:
    post:
      tags: [Auth & Sessions]
      summary: Échanger un code PIN valide contre la clé d'accès (appareil non authentifié)
      security: []
      responses:
        '200': { description: 'PIN consommé : { accessKey, handle }' }
        '404': { description: PIN invalide ou expiré }
        '429': { description: Trop de tentatives }
...
  /api/events:
    get:
      tags: [Notifications & Temps réel]
      summary: Flux SSE temps réel (notifications + messages)
      description: text/event-stream ; émet `ready`, `notif`, `message`, `ack`, `ping`.
      responses:
        '200': { description: Flux SSE, content: { text/event-stream: {} } }
        '401': { description: Non authentifié }
...
  /oauth/register:
    post:
      tags: [OAuth 2.1]
      summary: Enregistrement dynamique de client (RFC 7591)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [redirect_uris]
              properties:
                redirect_uris: { type: array, items: { type: string, format: uri } }
                client_name: { type: string }
                token_endpoint_auth_method: { type: string, enum: [none, client_secret_post] }
      responses:
        '201': { description: client_id (+ client_secret si confidentiel) }
        '400': { description: redirect_uris invalides }
  /oauth/authorize:
    get:
      tags: [OAuth 2.1]
      summary: Page de consentement (Authorization Code + PKCE S256)
      security: []
      parameters:
        - { name: response_type, in: query, required: true, schema: { type: string, enum: [code] } }
        - { name: client_id, in: query, required: true, schema: { type: string } }
        - { name: redirect_uri, in: query, required: true, schema: { type: string } }
        - { name: code_challenge, in: query, required: true, schema: { type: string } }
        - { name: code_challenge_method, in: query, required: true, schema: { type: string, enum: [S256] } }
        - { name: state, in: query, schema: { type: string } }
        - { name: scope, in: query, schema: { type: string, example: "mindlog:identity" } }
      responses:
        '200': { description: Page HTML de consentement (auth par clé d'accès) }
        '302': { description: Redirection avec erreur OAuth si paramètres invalides }
        '400': { description: client_id ou redirect_uri invalide }
    post:
      tags: [OAuth 2.1]
      summary: Décision de consentement → redirige avec un code d'autorisation
      security: []
      responses:
        '302': { description: Redirection vers redirect_uri avec code+state, ou error }
  /oauth/token:
    post:
      tags: [OAuth 2.1]
      summary: Échange de code / refresh contre des tokens
      security: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type: { type: string, enum: [authorization_code, refresh_token] }
                code: { type: string }
                code_verifier: { type: string }
                refresh_token: { type: string }
                redirect_uri: { type: string }
                client_id: { type: string }
                client_secret: { type: string }
      responses:
        '200': { description: "access_token, refresh_token, token_type, expires_in, scope" }
        '400': { description: invalid_grant / unsupported_grant_type }
        '401': { description: invalid_client }
...
  /mcp:
    post:
      tags: [MCP]
      summary: Connecteur MCP cloud (JSON-RPC 2.0, transport Streamable HTTP)
      description: |
        Endpoint MCP authentifié et **scopé à une seule identité**. Authentification :
        **token OAuth 2.1** (`Authorization: Bearer`, voie standard des connecteurs)
        ou, en repli, la **clé d'accès** brute (Bearer / `X-Access-Key` / `?key=`).
        Transport Streamable HTTP stateless (réponses JSON), 23 outils scopés.
        En 401, l'en-tête `WWW-Authenticate` pointe vers les métadonnées de ressource.
      security:
        - oauth2: []
        - accessKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Message JSON-RPC 2.0 (initialize, tools/list, tools/call…)
              properties:
                jsonrpc: { type: string, enum: ["2.0"] }
                id: { type: integer }
                method: { type: string, example: tools/list }
                params: { type: object }
      responses:
        '200': { description: Réponse JSON-RPC, content: { application/json: {} } }
        '401': { description: Clé d'accès requise ou invalide (erreur JSON-RPC -32001) }
        '413': { description: Corps trop volumineux }
        '429': { description: Trop de requêtes }
...